refactor(circular-deps): break ducks/metamask ↔ selectors/selectors cycle
Extract the base state-accessor selectors that the high-level
`ui/selectors/selectors.js` imports from `ui/ducks/metamask/metamask.js`
(`getConversionRate`, `isEIP1559Network`, `isNotEIP1559Network`,
`getLedgerTransportType`, `isAddressLedger`, `getIsUnlocked`, plus the
`findKeyringForAddress` helper used by `isAddressLedger`) into a new
`ui/ducks/metamask/base-selectors.js`.

`selectors.js` imports them from `base-selectors.js` directly so the duck
no longer appears in its dependency chain. `metamask.js` re-exports
them, preserving the public API for the duck's many other consumers.

No runtime change — pure file/import reorganization.
