+++ ui/ducks/metamask/metamask.js
@@ -15,17 +15,12 @@ import {
 import { KeyringType } from '../../../shared/constants/keyring';
 import { DEFAULT_AUTO_LOCK_TIME_LIMIT } from '../../../shared/constants/preferences';
 import { decGWEIToHexWEI } from '../../../shared/lib/conversion.utils';
-import { stripHexPrefix } from '../../../shared/lib/hexstring-utils';
-import { isEqualCaseInsensitive } from '../../../shared/lib/string-utils';
 import {
   accountsWithSendEtherInfoSelector,
   checkNetworkAndAccountSupports1559,
   getAddressBook,
 } from '../../selectors/selectors';
-import {
-  getProviderConfig,
-  getSelectedNetworkClientId,
-} from '../../../shared/lib/selectors/networks';
+import { getProviderConfig } from '../../../shared/lib/selectors/networks';
 import { getSelectedInternalAccount } from '../../../shared/lib/selectors/accounts';
 import * as actionConstants from '../../store/actionConstants';
 import { updateTransactionGasFees } from '../../store/actions';
@@ -316,13 +311,7 @@ export function getNativeCurrency(state) {
   return getProviderConfig(state).ticker;
 }
 
-export function getConversionRate(state) {
-  return (
-    getCurrencyRateControllerCurrencyRates(state)[
-      getProviderConfig(state).ticker
-    ]?.conversionRate ?? undefined
-  );
-}
+export { getConversionRate } from './base-selectors';
 
 export function getConversionRateByTicker(state, ticker) {
   return getCurrencyRateControllerCurrencyRates(state)[ticker]?.conversionRate;
@@ -340,34 +329,7 @@ export function getSendToAccounts(state) {
   return [...fromAccounts, ...addressBookAccounts];
 }
 
-/**
- * Function returns true if network details are fetched and it is found to not support EIP-1559
- *
- * @param state
- */
-export function isNotEIP1559Network(state) {
-  const selectedNetworkClientId = getSelectedNetworkClientId(state);
-  return (
-    state.metamask.networksMetadata[selectedNetworkClientId].EIPS[1559] ===
-    false
-  );
-}
-
-/**
- * Function returns true if network details are fetched and it is found to support EIP-1559
- *
- * @param state
- * @param networkClientId - The optional network client ID to check for EIP-1559 support. Defaults to the currently selected network.
- */
-export function isEIP1559Network(state, networkClientId) {
-  const selectedNetworkClientId = getSelectedNetworkClientId(state);
-
-  return (
-    state.metamask.networksMetadata?.[
-      networkClientId ?? selectedNetworkClientId
-    ]?.EIPS[1559] === true
-  );
-}
+export { isNotEIP1559Network, isEIP1559Network } from './base-selectors';
 
 function getGasFeeControllerEstimateType(state) {
   return state.metamask.gasEstimateType;
@@ -541,9 +503,7 @@ export function getIsWalletResetInProgress(state) {
   return state.metamask.isWalletResetInProgress;
 }
 
-export function getIsUnlocked(state) {
-  return state.metamask.isUnlocked;
-}
+export { getIsUnlocked } from './base-selectors';
 
 export function getSeedPhraseBackedUp(state) {
   return state.metamask.seedPhraseBackedUp;
@@ -576,48 +536,11 @@ export function getIsSeedlessPasswordOutdated(state) {
   return Boolean(state.metamask.passwordOutdatedCache?.isExpiredPwd);
 }
 
-/**
- * Given the redux state object and an address, finds a keyring that contains that address, if one exists
- *
- * @param {object} state - the redux state object
- * @param {string} address - the address to search for among the keyring addresses
- * @returns {object | undefined} The keyring which contains the passed address, or undefined
- */
-export function findKeyringForAddress(state, address) {
-  const keyring = state.metamask.keyrings.find((kr) => {
-    return kr.accounts.some((account) => {
-      return (
-        isEqualCaseInsensitive(account, addHexPrefix(address)) ||
-        isEqualCaseInsensitive(account, stripHexPrefix(address))
-      );
-    });
-  });
-
-  return keyring;
-}
-
-/**
- * Given the redux state object, returns the users preferred ledger transport type
- *
- * @param {object} state - the redux state object
- * @returns {string} The users preferred ledger transport type. One of 'webhid' on chrome or 'u2f' on firefox
- */
-export function getLedgerTransportType(state) {
-  return state.metamask.ledgerTransportType;
-}
-
-/**
- * Given the redux state object and an address, returns a boolean indicating whether the passed address is part of a Ledger keyring
- *
- * @param {object} state - the redux state object
- * @param {string} address - the address to search for among all keyring addresses
- * @returns {boolean} true if the passed address is part of a ledger keyring, and false otherwise
- */
-export function isAddressLedger(state, address) {
-  const keyring = findKeyringForAddress(state, address);
-
-  return keyring?.type === KeyringType.ledger;
-}
+export {
+  findKeyringForAddress,
+  getLedgerTransportType,
+  isAddressLedger,
+} from './base-selectors';
 
 /**
  * Given the redux state object, returns a boolean indicating whether the user has any Ledger accounts added to MetaMask (i.e. Ledger keyrings
